requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
patchright==1.44.0
requests==2.31.0
python-dotenv==1.0.1
aiohttp==3.9.5
sqlalchemy==2.0.30
python-multipart==0.0.9
httpx==0.27.0
